Don't add and remove a <span> to blink the cursor
Selecting text including the cursor was impossible, because the constant
change in the DOM removed the selection. This made copying lines
including the cursor tricky (you had 0.5s).

Remove all the blinking logic, but keep the <span>, so that blinking can
be done with css animations. That enables slicker blinking and consumes
less resources.