Remove broken refresh optimization
This optimization temporarily removed the terminal element from the DOM
while updating lines. It only did this when the number of lines was big
enough. This breaks focus handling though: if the terminal had the
focus, it lost it when the element was removed.

In my testing, this doesn't degrade performance in normal use-cases.