New option added: disableScrollingRegion
New `disableScrollingRegion` option is added to
the Terminal constructor.

When set to `true`, the Control Sequence Introducer `r`
is ignored.

Useful when Terminal is connected to a serial console
of a virtual machine which is being restarted. In that
case, the `CSI r` is received with incorrect value
causing Terminal to fail.

When the option is not explicitely set, behavior
remains same.